Pushbutton lib: improved debouncing logic and reduced width of PrevTimeMillis vars. I'm hesitant to make them 1 byte because that means the functions would wrongly think 15 ms hasn't elapsed yet if the function is called after e.g. 266 ms.

+++ Pushbutton/Pushbutton.h
@@ -14,33 +14,33 @@
 class Pushbutton
 {
   public:
-    
+
     // constructor; takes arguments specifying whether to enable internal pull-up
     // and the default state of the pin that the button is connected to
     Pushbutton(unsigned char pin, unsigned char pullUp = PULL_UP_ENABLED, unsigned char defaultState = DEFAULT_STATE_HIGH);
-    
+
     // wait for button to be pressed, released, or pressed and released
     void waitForPress();
     void waitForRelease();
     void waitForButton();
-    
+
     // indicates whether button is currently pressed
     boolean isPressed();
-    
+
     // more complex functions that return true once for each button transition
     // from released to pressed or pressed to released
     boolean getSingleDebouncedPress();
     boolean getSingleDebouncedRelease();
-    
+
   private:
-  
+
     unsigned char _pin;
     unsigned char _pullUp;
     unsigned char _defaultState;
     unsigned char gsdpState;
     unsigned char gsdrState;
-    unsigned long gsdpPrevTimeMillis;
-    unsigned long gsdrPrevTimeMillis;
+    unsigned int gsdpPrevTimeMillis;
+    unsigned int gsdrPrevTimeMillis;
     boolean initialized;
 
     inline void init()
@@ -51,10 +51,10 @@ class Pushbutton
         init2();
       }
     }
-    
+
     // initializes I/O pin for use as button input
     void init2();
-    
+
     boolean _isPressed();
 };
